Hydroid Solr Mapper
New component to remove specific references to the hydroid vocabulary
from the enhancer service. The idea is that different vocabularies would
have their own mapper to convert to Solr while the enhancer service
remains generic.
